refactor thread stack code
