+++ SVA/lib/init.c
@@ -84,6 +84,7 @@
 #include "sva/config.h"
 #include "sva/state.h"
 #include "sva/interrupt.h"
+#include "thread_stack.h"
 
 #include <string.h>
 #include <limits.h>
@@ -483,7 +484,6 @@ sva_init_primary () {
   /* Initialize the processor ID */
   init_procID();
 
-  extern void init_threads(void);
   init_threads();
 
   /* Initialize the IDT of the primary processor */
