+++ SVA/lib/interrupt.c
@@ -18,6 +18,7 @@
 #include "sva/interrupt.h"
 #include "sva/state.h"
 #include "sva/keys.h"
+#include "thread_stack.h"
 
 /* Debug flags for printing data */
 #define DEBUG       0
@@ -30,81 +31,6 @@
 #define sva_safemem_exception   (27)
 
 extern void * interrupt_table[256];
-
-/* Definitions of functions and variables related to the stack 
-   that is used for finding the next free thread.
-*/
-#define THREAD_SIZE 10000
-#define spin_lock(l)  while (__sync_lock_test_and_set((l), 1))
-#define spin_unlock(l) __sync_lock_release((l))
-#define init_lock(l) *(l) = 0
-#define NULL  0
-
-/* Pre-allocate a large number of SVA Threads */
-static struct SVAThread realThreads[4096] __attribute__ ((aligned (16)))
-__attribute__ ((section ("svamem")));
-
-struct SVAThread *Threads = realThreads;
-
-typedef volatile int lock_t;
-
-/* Stack for storing the free treads */
-struct FT_stack{
-  struct SVAThread *threads[THREAD_SIZE];
-  int top;
-  int initialized;
-  lock_t lock;
-};
-
-/* Initialization of the stack */
-static struct FT_stack fthreads __attribute__ ((aligned (16))) __attribute__ ((section ("svamem"))) = {
-  .top = -1,
-  .lock = 0,
-  .initialized = 0,
-};
-
-/* Push function for the stack. The action has to be protected by a spinlock. */
-void ftstack_push(struct SVAThread *thread){
-  spin_lock(&fthreads.lock);
-
-  fthreads.top++;
-  fthreads.threads[fthreads.top] = thread;
-
-  spin_unlock(&fthreads.lock);
-}
-
-/* Pop function for the stack. */
-struct SVAThread *ftstack_pop(void) {
-
-  struct SVAThread *result = NULL;
-
-  spin_lock(&fthreads.lock);
-
-  if(!fthreads.initialized) {
-
-    for (unsigned index = 0; index < 4096; ++index) {
-      Threads[index].used = 0;
-      fthreads.top++;
-      fthreads.threads[fthreads.top] = &Threads[index];
-    }
-
-    fthreads.initialized = 1;
-  }
-
-  if(fthreads.top == -1) {
-    spin_unlock(&fthreads.lock);
-    panic("");
-    return NULL;
-  }
-  
-  result = fthreads.threads[fthreads.top];
-  fthreads.top--;
-
-  spin_unlock(&fthreads.lock);
-
-  return result;
-}
-
 
 
 
@@ -155,97 +81,6 @@ static struct CPUState realCPUState[numProcessors] __attribute__((aligned(16)))
 __attribute__ ((section ("svamem")));
 struct CPUState * CPUState = realCPUState;
 
-
-void
-init_threads(void) {
-  return;
-}
-
-/*
- * Function: randomNumber()
- *
- * Description:
- *  Use the rdrand instruction to generate a 64-bit random number.
- */
-static inline uintptr_t
-randomNumber (void) {
-  uintptr_t rand;
-  __asm__ __volatile__ ("1: rdrand %0\n"
-                        "jae 1\n" : "=r" (rand));
-  return rand;
-}
-
-/*
- * Function: findNextFreeThread()
- *
- * Description:
- *  This function returns the index of the next thread which is not in use.
- */
-struct SVAThread *
-findNextFreeThread (void) {
-  /*
-   * Find the next free thread.
-   */
-  struct SVAThread *newThread = ftstack_pop();
-  if (newThread) {
-    /*
-     * Do some basic initialization of the thread.
-     */
-    newThread->integerState.valid = 0;
-    newThread->savedICIndex = 0;
-    newThread->ICFPIndex = 0;
-    newThread->secmemSize = 0;
-    newThread->numPushTargets = 0;
-    newThread->secmemPML4e = 0;
-    newThread->isInitialForCPU = 0;
-
-    /* 
-     * This function currently sets the thread secret with a default
-     * statically defined key.  However, in the future will obtain said key
-     * from the executable image.  There is also some issue with
-     * bootstrapping the initial key and whether or not on the first
-     * execution of an application the key will need to be generated by SVA.
-     * Thus, future design is yet to be done, however, the following function
-     * should suffice to enable any of the above scenarios.
-     *
-     * TODO: The function currently uses a dummy static key, but in the
-     * future will obtain the key from the executable image and then
-     * decrypted with the VirtualGhost private key. 
-     */
-    if (vg) {
-      extern void init_thread_key (struct SVAThread * thread);
-      init_thread_key(newThread);
-    }
-
-#if DEBUG
-    printf("<<<< SVA: Created new private key: value: %s\n",
-            newThread->secret.key); 
-#endif
-
-    /*
-     * Use the next-to-last interrupt context in the list as the first
-     * interrupt context.  This may be slightly wasteful, but it's a little
-     * easier to make it work correctly right now.
-     *
-     * The processor's IST3 field should be configured so that the next
-     * interrupt context is at maxIC - 2.
-     */
-    sva_icontext_t * icontextp = newThread->interruptContexts + maxIC - 1;
-    newThread->integerState.ist3 = ((uintptr_t) icontextp) - 0x10;
-    newThread->integerState.kstackp = newThread->integerState.ist3; 
-
-    /*
-     * Generate a random identifier for the new thread.
-     */
-    if (vg) {
-      newThread->rid = randomNumber();
-    }
-    return newThread;
-  }
-
-  panic ("SVA: findNextFreeThread: Exhausted SVA Threads!\n");
-  return 0;
-}
 
 /*
  * Intrinsic: sva_getCPUState()
