+++ SVA/lib/state.c
@@ -25,6 +25,7 @@
 #include <sva/mmu.h>
 #include "sva/mmu_intrinsics.h"
 #include <sva/x86.h>
+#include "thread_stack.h"
 
 /*****************************************************************************
  * Externally Visibile Utility Functions
@@ -1057,7 +1058,6 @@ sva_release_stack (uintptr_t id) {
   newThread->used = 0;
 
   /* Push the thread into the stack of free threads since it can be reused */
-  extern void ftstack_push(struct SVAThread *thread);
   ftstack_push(newThread);
 
   return;
@@ -1159,7 +1159,6 @@ sva_init_stack (unsigned char * start_stackp,
   /*
    * Allocate a new SVA thread.
    */
-  extern struct SVAThread * findNextFreeThread (void);
   struct SVAThread * newThread = findNextFreeThread();
